+++ src/main/java/org/example/oddventure/domain/admin/controller/AdminMatchController.java
@@ -1,0 +1,58 @@
+package org.example.oddventure.domain.admin.controller;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.example.oddventure.common.dto.response.ApiResponse;
+import org.example.oddventure.domain.admin.dto.request.MatchCreateRequest;
+import org.example.oddventure.domain.admin.dto.request.MatchUpdateRequest;
+import org.example.oddventure.domain.admin.dto.response.MatchAdminResponse;
+import org.example.oddventure.domain.admin.service.AdminMatchService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/admin/matches")
+public class AdminMatchController {
+
+    private final AdminMatchService matchService;
+    private final AdminMatchService adminMatchService;
+
+    // 매치 생성
+    @PostMapping("/matches")
+    public ResponseEntity<ApiResponse<MatchAdminResponse>> createMatch(
+            @Valid @RequestBody MatchCreateRequest request
+    ) {
+        MatchAdminResponse response = adminMatchService.createMatch(request);
+        return ApiResponse.created(response, "매치가 생성되었습니다.");
+    }
+
+    // 매치 상태 수정
+    @PatchMapping("/matches/{matchId}")
+    public ResponseEntity<ApiResponse<MatchAdminResponse>> updateMatchStatus(
+            @PathVariable Long matchId,
+            @Valid @RequestBody MatchUpdateRequest request
+    ) {
+        MatchAdminResponse response = adminMatchService.updateMatch(matchId, request);
+        return ApiResponse.success(response, "매치 정보가 수정되었습니다.");
+    }
+
+    // 매치 일정 연동
+    @PostMapping("/fetch")
+    public ResponseEntity<ApiResponse<Void>> fetchMatches() {
+        adminMatchService.fetchMatches();
+        return ApiResponse.success("매치가 연동되었습니다.");
+    }
+
+    // 매치 결과 연동
+    @PatchMapping("/fetch/{fetchId}")
+    public ResponseEntity<ApiResponse<Void>> syncMatchResult(@PathVariable Long fetchId) {
+        adminMatchService.fetchMatchResult(fetchId);
+        return ApiResponse.success("매치 결과가 연동되었습니다.");
+    }
+}
